tools/nolibc: merge i386 and x86_64 into a single x86 arch
This remained the only exception to the kernel's architectures
organization and it's always a bit cumbersome to deal with. Let's merge
i386 and x86_64 into x86. This will result in a single arch-x86.h file
by default, and we'll no longer need to merge the two manually during
installation. Requesting either i386 or x86_64 will also result in
installing x86.

Signed-off-by: Willy Tarreau <w@1wt.eu>